Refactor FarmingPatch
